WIP Centering Logo - Refactoring mobile and desktop headers
